Split tests into several files
