+++ tests/tests-select.js
@@ -2,72 +2,7 @@
 "use strict";
 
 var expect = require('chai').expect;
-var m = require('../simpleSqlParser');
-
-function ok(ast) {
-	return {
-		status: true,
-		value: ast
-	};
-}
-
-function testAst(comment, query, ast) {
-	expect(m.sql2ast(query)).to.eql(ok(ast), comment + ': ' + query);
-}
-
-function testBackAndForth(comment, query) {
-	expect(m.ast2sql(m.sql2ast(query))).to.eql(query, comment + ': ' + query);
-}
-
-function isArray(variable, message) {
-	expect(variable).to.be.instanceOf(Array, message);
-}
-
-function isObject(variable, message) {
-	expect(variable).not.to.be.instanceOf(Array, message);
-}
-
-test('sql2ast - API', function() {
-
-	var q = [
-		'SELECT * FROM table',
-		'INSERT INTO table VALUES (1, 2, 3)',
-		'UPDATE table SET col = "value"',
-		'DELETE FROM table',
-	];
-	var ast = q.map(m.sql2ast);
-
-	var types = ['select', 'delete', 'insert', 'update'];
-
-	ast.forEach(function(a) {
-		expect(a.status).to.equal(true, 'Parser must parse valid SQL');
-		expect(types.indexOf(a.value.type)).not.to.equal(-1, 'AST must contain a valid type');
-
-		if (a.value.type === types[0]) {
-			isArray(a.value.select, "(SELECT) AST must contain a 'select' array");
-			isArray(a.value.from, "(SELECT) AST must contain a 'from' array");
-			isArray(a.value.join, "(SELECT) AST must contain a 'join' array");
-			isObject(a.value.where, "(SELECT) AST must contain a 'where' object");
-			isArray(a.value.group, "(SELECT) AST must contain a 'group' array");
-			isArray(a.value.order, "(SELECT) AST must contain a 'order' array");
-			isObject(a.value.limit, "(SELECT) AST must contain a 'limit' object");
-		}
-		else if (a.value.type === types[1]) {
-			isArray(a.value.from, "(DELETE) AST must contain a 'from' array");
-			isObject(a.value.where, "(DELETE) AST must contain a 'where' object");
-		}
-		else if (a.value.type === types[2]) {
-			isObject(a.value.into, "(INSERT) AST must contain a 'into' object");
-			isArray(a.value.values, "(INSERT) AST must contain a 'values' array");
-		}
-		else if (a.value.type === types[3]) {
-			isObject(a.value.table, "(UPDATE) AST must contain a 'table' object");
-			isArray(a.value.values, "(UPDATE) AST must contain a 'values' array");
-			isObject(a.value.where, "(UPDATE) AST must contain a 'where' object");
-		}
-	});
-
-});
+var h = require('./helpers.js');
 
 var Select = [
 	{
@@ -465,211 +400,14 @@ var Select = [
 	},
 ];
 
-var Insert = [
-	{
-		c: 'Simple insert',
-		q: 'INSERT INTO table VALUES (1, 2, 3)',
-		a: {
-			type: 'insert',
-			into: {
-				expression: 'table',
-				table: 'table',
-				alias: null,
-			},
-			values: [
-				{ target: null, value: '1'},
-				{ target: null, value: '2'},
-				{ target: null, value: '3'},
-			],
-		},
-	},
-	{
-		c: 'Complex values',
-		q: 'INSERT INTO table VALUES (1 + 9, FUNC(2, col), "string")',
-		a: {
-			type: 'insert',
-			into: {
-				expression: 'table',
-				table: 'table',
-				alias: null,
-			},
-			values: [
-				{ target: null, value: '1 + 9'},
-				{ target: null, value: 'FUNC(2, col)'},
-				{ target: null, value: '"string"'},
-			],
-		},
-	},
-	{
-		c: 'Insert with columns',
-		q: 'INSERT INTO table (col1, `col2`, col3) VALUES (1, 2, 3)',
-		a: {
-			type: 'insert',
-			into: {
-				expression: 'table',
-				table: 'table',
-				alias: null,
-			},
-			values: [
-				{ target: { expression: 'col1', column: 'col1' }, value: '1'},
-				{ target: { expression: '`col2`', column: 'col2' }, value: '2'},
-				{ target: { expression: 'col3', column: 'col3' }, value: '3'},
-			],
-		},
-	},
-];
-
-var Update = [
-	{
-		c: 'Simple update',
-		q: 'UPDATE table SET col = "value"',
-		a: {
-			type: 'update',
-			table: {
-				expression: 'table',
-				table: 'table',
-				alias: null,
-			},
-			where: null,
-			values: [
-				{ target: { expression: 'col', column: 'col' }, value: '"value"'},
-			],
-		},
-	},
-	{
-		c: 'Several columns',
-		q: 'UPDATE table SET col = "value", col2 = NULL, table.col3 = col',
-		a: {
-			type: 'update',
-			table: {
-				expression: 'table',
-				table: 'table',
-				alias: null,
-			},
-			where: null,
-			values: [
-				{ target: { expression: 'col', column: 'col' }, value: '"value"'},
-				{ target: { expression: 'col2', column: 'col2' }, value: 'NULL'},
-				{ target: { expression: 'table.col3', column: 'col3' }, value: 'col'},
-			],
-		},
-	},
-	{
-		c: 'Where #1',
-		q: 'UPDATE table SET col = "value" WHERE this >= that AND col IS NOT NULL',
-		a: {
-			type: 'update',
-			table: {
-				expression: 'table',
-				table: 'table',
-				alias: null,
-			},
-			where: {
-				expression: "this >= that AND col IS NOT NULL",
-			},
-			values: [
-				{ target: { expression: 'col', column: 'col' }, value: '"value"'},
-			],
-		},
-	},
-];
-
-var Delete = [
-	{
-		c: 'Simple delete',
-		q: 'DELETE FROM table',
-		a: {
-			type: 'delete',
-			from: [
-				{ expression: 'table', table: 'table', alias: null },
-			],
-			where: null,
-		},
-	},
-	{
-		c: 'Several tables with aliases',
-		q: 'DELETE FROM table1 AS t1, table2 "t2"',
-		a: {
-			type: 'delete',
-			from: [
-				{ expression: 'table1 AS t1', table: 'table1', alias: 't1' },
-				{ expression: 'table2 "t2"', table: 'table2', alias: 't2' },
-			],
-			where: null,
-		},
-	},
-	{
-		c: 'Where #1',
-		q: 'DELETE FROM table WHERE this >= that AND col IS NOT NULL',
-		a: {
-			type: 'delete',
-			from: [
-				{ expression: 'table', table: 'table', alias: null },
-			],
-			where: {
-				expression: "this >= that AND col IS NOT NULL",
-			},
-		},
-	},
-	{
-		c: 'Where #2',
-		q: 'DELETE FROM table WHERE (FUNC(this) = "string") AND (1+5 OR col1)',
-		a: {
-			type: 'delete',
-			from: [
-				{ expression: 'table', table: 'table', alias: null },
-			],
-			where: {
-				expression: "(FUNC(this) = \"string\") AND (1+5 OR col1)",
-			},
-		},
-	},
-];
-
 test('sql2ast - select', function() {
 	Select.forEach(function(test) {
-		testAst(test.c, test.q, test.a);
-	});
-});
-
-test('sql2ast - insert', function() {
-	Insert.forEach(function(test) {
-		testAst(test.c, test.q, test.a);
-	});
-});
-
-test('sql2ast - update', function() {
-	Update.forEach(function(test) {
-		testAst(test.c, test.q, test.a);
-	});
-});
-
-test('sql2ast - delete', function() {
-	Delete.forEach(function(test) {
-		testAst(test.c, test.q, test.a);
+		h.testAst(test.c, test.q, test.a);
 	});
 });
 
 test('ast2sql - select', function() {
 	Select.forEach(function(test) {
-		testBackAndForth(test.c, test.q);
-	});
-});
-
-test('ast2sql - insert', function() {
-	Insert.forEach(function(test) {
-		testBackAndForth(test.c, test.q);
-	});
-});
-
-test('ast2sql - update', function() {
-	Update.forEach(function(test) {
-		testBackAndForth(test.c, test.q);
-	});
-});
-
-test('ast2sql - delete', function() {
-	Delete.forEach(function(test) {
-		testBackAndForth(test.c, test.q);
+		h.testBackAndForth(test.c, test.q);
 	});
 });
